move HttpClientInitializer
