+++ http-client/src/main/java/io/micronaut/http/client/netty/ConnectionManager.java
@@ -8,9 +8,13 @@ import io.micronaut.http.client.HttpClientConfiguration;
 import io.micronaut.http.client.exceptions.HttpClientException;
 import io.micronaut.http.netty.channel.ChannelPipelineCustomizer;
 import io.micronaut.http.netty.channel.ChannelPipelineListener;
+import io.micronaut.http.netty.stream.DefaultHttp2Content;
+import io.micronaut.http.netty.stream.Http2Content;
+import io.micronaut.http.netty.stream.HttpStreamsClientHandler;
 import io.micronaut.scheduling.instrument.InvocationInstrumenter;
 import io.micronaut.websocket.exceptions.WebSocketSessionException;
 import io.netty.bootstrap.Bootstrap;
+import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFactory;
 import io.netty.channel.ChannelFuture;
@@ -18,6 +22,7 @@ import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
+import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.EventLoopGroup;
@@ -28,11 +33,32 @@ import io.netty.channel.pool.ChannelPool;
 import io.netty.channel.pool.ChannelPoolMap;
 import io.netty.channel.pool.FixedChannelPool;
 import io.netty.channel.pool.SimpleChannelPool;
+import io.netty.channel.socket.SocketChannel;
+import io.netty.handler.codec.LineBasedFrameDecoder;
+import io.netty.handler.codec.http.DefaultHttpContent;
+import io.netty.handler.codec.http.FullHttpMessage;
+import io.netty.handler.codec.http.HttpClientCodec;
+import io.netty.handler.codec.http.HttpContent;
+import io.netty.handler.codec.http.HttpContentDecompressor;
+import io.netty.handler.codec.http.HttpMessage;
+import io.netty.handler.codec.http.HttpObjectAggregator;
+import io.netty.handler.codec.http.HttpUtil;
+import io.netty.handler.codec.http.LastHttpContent;
 import io.netty.handler.codec.http.websocketx.extensions.compression.WebSocketClientCompressionHandler;
+import io.netty.handler.codec.http2.DefaultHttp2Connection;
+import io.netty.handler.codec.http2.Http2Connection;
+import io.netty.handler.codec.http2.Http2FrameLogger;
+import io.netty.handler.codec.http2.Http2Stream;
+import io.netty.handler.codec.http2.HttpToHttp2ConnectionHandler;
+import io.netty.handler.codec.http2.HttpToHttp2ConnectionHandlerBuilder;
+import io.netty.handler.logging.LoggingHandler;
 import io.netty.handler.ssl.SslContext;
+import io.netty.handler.ssl.SslHandler;
+import io.netty.handler.timeout.IdleStateEvent;
 import io.netty.handler.timeout.IdleStateHandler;
 import io.netty.handler.timeout.ReadTimeoutHandler;
 import io.netty.resolver.NoopAddressResolverGroup;
+import io.netty.util.Attribute;
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.Promise;
 import org.slf4j.Logger;
@@ -170,7 +196,7 @@ final class ConnectionManager {
         int port = requestKey.getPort();
         Bootstrap localBootstrap = bootstrap.clone();
         initBootstrapForProxy(localBootstrap, sslCtx != null, host, port);
-        localBootstrap.handler(httpClient.new HttpClientInitializer(
+        localBootstrap.handler(new HttpClientInitializer(
             sslCtx,
             host,
             port,
@@ -318,7 +344,7 @@ final class ConnectionManager {
 
         bootstrap.remoteAddress(requestKey.getHost(), requestKey.getPort());
         initBootstrapForProxy(bootstrap, sslContext != null, requestKey.getHost(), requestKey.getPort());
-        bootstrap.handler(httpClient.new HttpClientInitializer(
+        bootstrap.handler(new HttpClientInitializer(
             sslContext,
             requestKey.getHost(),
             requestKey.getPort(),
@@ -405,7 +431,7 @@ final class ConnectionManager {
                 };
                 ch.pipeline().addLast(failureHandler);
 
-                ch.pipeline().addLast(ChannelPipelineCustomizer.HANDLER_HTTP_CLIENT_INIT, httpClient.new HttpClientInitializer(
+                ch.pipeline().addLast(ChannelPipelineCustomizer.HANDLER_HTTP_CLIENT_INIT, new HttpClientInitializer(
                     key.isSecure() ? sslContext : null,
                     key.getHost(),
                     key.getPort(),
@@ -472,6 +498,253 @@ final class ConnectionManager {
                 }
             }
         };
+    }
+
+    /**
+     * Initializes the HTTP client channel.
+     */
+    protected class HttpClientInitializer extends ChannelInitializer<SocketChannel> {
+
+        final SslContext sslContext;
+        final String host;
+        final int port;
+        final boolean stream;
+        final boolean proxy;
+        final boolean acceptsEvents;
+        DefaultHttpClient.Http2SettingsHandler settingsHandler;
+        final Consumer<ChannelHandlerContext> contextConsumer;
+        private NettyClientCustomizer channelCustomizer;
+
+        /**
+         * @param sslContext      The ssl context
+         * @param host            The host
+         * @param port            The port
+         * @param stream          Whether is stream
+         * @param proxy           Is this a streaming proxy
+         * @param acceptsEvents   Whether an event stream is accepted
+         * @param contextConsumer The context consumer
+         */
+        protected HttpClientInitializer(SslContext sslContext,
+                                        String host,
+                                        int port,
+                                        boolean stream,
+                                        boolean proxy,
+                                        boolean acceptsEvents,
+                                        Consumer<ChannelHandlerContext> contextConsumer) {
+            this.sslContext = sslContext;
+            this.stream = stream;
+            this.host = host;
+            this.port = port;
+            this.proxy = proxy;
+            this.acceptsEvents = acceptsEvents;
+            this.contextConsumer = contextConsumer;
+        }
+
+        /**
+         * @param ch The channel
+         */
+        @Override
+        protected void initChannel(SocketChannel ch) {
+            channelCustomizer = clientCustomizer.specializeForChannel(ch, NettyClientCustomizer.ChannelRole.CONNECTION);
+            ch.attr(DefaultHttpClient.CHANNEL_CUSTOMIZER_KEY).set(channelCustomizer);
+
+            ChannelPipeline p = ch.pipeline();
+
+            Proxy proxy = configuration.resolveProxy(sslContext != null, host, port);
+            if (!Proxy.NO_PROXY.equals(proxy)) {
+                httpClient.configureProxy(p, proxy);
+            }
+
+            if (httpVersion == HttpVersion.HTTP_2_0) {
+                final Http2Connection connection = new DefaultHttp2Connection(false);
+                final HttpToHttp2ConnectionHandlerBuilder builder =
+                        httpClient.newHttp2ConnectionHandlerBuilder(connection, configuration, stream);
+
+                configuration.getLogLevel().ifPresent(logLevel -> {
+                    try {
+                        final io.netty.handler.logging.LogLevel nettyLevel = io.netty.handler.logging.LogLevel.valueOf(
+                                logLevel.name()
+                        );
+                        builder.frameLogger(new Http2FrameLogger(nettyLevel, DefaultHttpClient.class));
+                    } catch (IllegalArgumentException e) {
+                        throw httpClient.customizeException(new HttpClientException("Unsupported log level: " + logLevel));
+                    }
+                });
+                HttpToHttp2ConnectionHandler connectionHandler = builder
+                        .build();
+                if (sslContext != null) {
+                    httpClient.configureHttp2Ssl(this, ch, sslContext, host, port, connectionHandler);
+                } else {
+                    httpClient.configureHttp2ClearText(this, ch, connectionHandler);
+                }
+                channelCustomizer.onInitialPipelineBuilt();
+            } else {
+                if (stream) {
+                    // for streaming responses we disable auto read
+                    // so that the consumer is in charge of back pressure
+                    ch.config().setAutoRead(false);
+                }
+
+                configuration.getLogLevel().ifPresent(logLevel -> {
+                    try {
+                        final io.netty.handler.logging.LogLevel nettyLevel = io.netty.handler.logging.LogLevel.valueOf(
+                                logLevel.name()
+                        );
+                        p.addLast(new LoggingHandler(DefaultHttpClient.class, nettyLevel));
+                    } catch (IllegalArgumentException e) {
+                        throw httpClient.customizeException(new HttpClientException("Unsupported log level: " + logLevel));
+                    }
+                });
+
+                if (sslContext != null) {
+                    SslHandler sslHandler = sslContext.newHandler(ch.alloc(), host, port);
+                    sslHandler.setHandshakeTimeoutMillis(configuration.getSslConfiguration().getHandshakeTimeout().toMillis());
+                    p.addLast(ChannelPipelineCustomizer.HANDLER_SSL, sslHandler);
+                }
+
+                // Pool connections require alternative timeout handling
+                if (poolMap == null) {
+                    // read timeout settings are not applied to streamed requests.
+                    // instead idle timeout settings are applied.
+                    if (stream) {
+                        Optional<Duration> readIdleTime = configuration.getReadIdleTimeout();
+                        if (readIdleTime.isPresent()) {
+                            Duration duration = readIdleTime.get();
+                            if (!duration.isNegative()) {
+                                p.addLast(HANDLER_IDLE_STATE, new IdleStateHandler(
+                                        duration.toMillis(),
+                                        duration.toMillis(),
+                                        duration.toMillis(),
+                                        TimeUnit.MILLISECONDS
+                                ));
+                            }
+                        }
+                    }
+                }
+
+                addHttp1Handlers(p);
+                channelCustomizer.onInitialPipelineBuilt();
+                onStreamPipelineBuilt();
+            }
+        }
+
+        /**
+         * Called when the stream pipeline is fully set up (all handshakes completed) and we can
+         * start processing requests.
+         */
+        void onStreamPipelineBuilt() {
+            channelCustomizer.onStreamPipelineBuilt();
+        }
+
+        void addHttp1Handlers(ChannelPipeline p) {
+            p.addLast(ChannelPipelineCustomizer.HANDLER_HTTP_CLIENT_CODEC, new HttpClientCodec());
+
+            p.addLast(ChannelPipelineCustomizer.HANDLER_HTTP_DECODER, new HttpContentDecompressor());
+
+            int maxContentLength = configuration.getMaxContentLength();
+
+            if (!stream) {
+                p.addLast(ChannelPipelineCustomizer.HANDLER_HTTP_AGGREGATOR, new HttpObjectAggregator(maxContentLength) {
+                    @Override
+                    protected void finishAggregation(FullHttpMessage aggregated) throws Exception {
+                        if (!HttpUtil.isContentLengthSet(aggregated)) {
+                            if (aggregated.content().readableBytes() > 0) {
+                                super.finishAggregation(aggregated);
+                            }
+                        }
+                    }
+                });
+            }
+            addEventStreamHandlerIfNecessary(p);
+            addFinalHandler(p);
+            for (ChannelPipelineListener pipelineListener : httpClient.connectionManager.pipelineListeners) {
+                pipelineListener.onConnect(p);
+            }
+        }
+
+        void addEventStreamHandlerIfNecessary(ChannelPipeline p) {
+            // if the content type is a SSE event stream we add a decoder
+            // to delimit the content by lines (unless we are proxying the stream)
+            if (acceptsEventStream() && !proxy) {
+                p.addLast(ChannelPipelineCustomizer.HANDLER_MICRONAUT_SSE_EVENT_STREAM, new LineBasedFrameDecoder(configuration.getMaxContentLength(), true, true) {
+
+                    @Override
+                    public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
+                        if (msg instanceof HttpContent) {
+                            if (msg instanceof LastHttpContent) {
+                                super.channelRead(ctx, msg);
+                            } else {
+                                Attribute<Http2Stream> streamKey = ctx.channel().attr(DefaultHttpClient.STREAM_KEY);
+                                if (msg instanceof Http2Content) {
+                                    streamKey.set(((Http2Content) msg).stream());
+                                }
+                                try {
+                                    super.channelRead(ctx, ((HttpContent) msg).content());
+                                } finally {
+                                    streamKey.set(null);
+                                }
+                            }
+                        } else {
+                            super.channelRead(ctx, msg);
+                        }
+                    }
+                });
+
+                p.addLast(ChannelPipelineCustomizer.HANDLER_MICRONAUT_SSE_CONTENT, new SimpleChannelInboundHandlerInstrumented<ByteBuf>(httpClient.connectionManager.instrumenter, false) {
+
+                    @Override
+                    public boolean acceptInboundMessage(Object msg) {
+                        return msg instanceof ByteBuf;
+                    }
+
+                    @Override
+                    protected void channelReadInstrumented(ChannelHandlerContext ctx, ByteBuf msg) {
+                        try {
+                            Attribute<Http2Stream> streamKey = ctx.channel().attr(DefaultHttpClient.STREAM_KEY);
+                            Http2Stream http2Stream = streamKey.get();
+                            if (http2Stream != null) {
+                                ctx.fireChannelRead(new DefaultHttp2Content(msg.copy(), http2Stream));
+                            } else {
+                                ctx.fireChannelRead(new DefaultHttpContent(msg.copy()));
+                            }
+                        } finally {
+                            msg.release();
+                        }
+                    }
+                });
+
+            }
+        }
+
+        /**
+         * Allows overriding the final handler added to the pipeline.
+         *
+         * @param pipeline The pipeline
+         */
+        protected void addFinalHandler(ChannelPipeline pipeline) {
+            pipeline.addLast(
+                    ChannelPipelineCustomizer.HANDLER_HTTP_STREAM,
+                    new HttpStreamsClientHandler() {
+                @Override
+                public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
+                    if (evt instanceof IdleStateEvent) {
+                        // close the connection if it is idle for too long
+                        ctx.close();
+                    }
+                    super.userEventTriggered(ctx, evt);
+                }
+
+                @Override
+                protected boolean isValidInMessage(Object msg) {
+                    // ignore data on stream 1, that is the response to our initial upgrade request
+                    return super.isValidInMessage(msg) && (sslContext != null || !httpClient.discardH2cStream((HttpMessage) msg));
+                }
+            });
+        }
+
+        private boolean acceptsEventStream() {
+            return this.acceptsEvents;
+        }
     }
 
     class PoolHandle {
